requirements.txt
openai-whisper>=20240930
torch
tqdm
colorama
